+++ mmm_eval/core/constants.py
@@ -4,11 +4,10 @@
 class ValidationTestConstants:
     """Constants for the validation tests."""
 
-    #TRAIN_TEST_SPLIT_TEST_PROPORTION = 0.2
     ACCURACY_TEST_SIZE = 8
     RANDOM_STATE = 42
-    N_SPLITS = 4
-    TIME_SERIES_CROSS_VALIDATION_TEST_SIZE = 4  # 4 representing a 4 week refresh
+    N_SPLITS = 5
+    TIME_SERIES_CROSS_VALIDATION_TEST_SIZE = 4
 
     class PerturbationConstants:
         """Constants for the perturbation test."""